iframe-proxy: share the proxy across VS Code and standalone via lib
Move the host-agnostic proxy into lib/src/host as one TypeScript source so
both hosts run the same code instead of reimplementing it:

  - iframe-proxy-rewrite.ts — pure policy/rewriting (shim, instrumentHtml,
    refusesFraming/hasRestrictiveFrameAncestors, loopback/SSRF checks, error
    pages). Now unit-tested (17 cases), incl. the frame-ancestors logic.
  - iframe-proxy.ts — the Node http/net server, logger injected so it depends
    on neither host.
  - IframeProxyResult moves to a dependency-free leaf (platform/
    iframe-proxy-types.ts, re-exported) so the Node code doesn't pull the
    browser type-graph into a Node compile.

VS Code: iframe-proxy-host.ts collapses to a thin wrapper that injects `log`.

Standalone: build-sidecar-proxy.mjs esbuilds the shared TS into
sidecar/iframe-proxy.cjs (wired into stage/build/tauri, gitignored); the
sidecar handles iframe:createProxyUrl over stdio; a Rust
iframe_create_proxy_url command bridges it; tauri-adapter implements
createIframeProxyUrl; and the Tauri CSP gains frame-src for the loopback
proxy. IframePanel is shared, so the webview needs no changes.

Verified: lib tsc + 609 tests, VS Code bundle, sidecar cjs builds and
runtime-proxies, standalone tsc. The Rust command needs a machine with the
Rust toolchain to compile/run (cargo absent here).
