+++ lib/src/host/iframe-proxy-rewrite.ts
@@ -1,0 +1,153 @@
+/**
+ * Pure, dependency-free helpers for the iframe transparent proxy
+ * (docs/specs/dor-iframe.md → "The Transparent Proxy").
+ *
+ * Split out from the Node server (`iframe-proxy.ts`) so the policy/rewriting
+ * logic is shared by every host that runs the proxy (VS Code extension host,
+ * Tauri sidecar) and is unit-testable without standing up a server. Nothing
+ * here imports `http`/`net`; headers are typed structurally so this file is
+ * runtime-agnostic.
+ */
+
+/** Header bag shape both `http.IncomingHttpHeaders` and a plain map satisfy. */
+export type ProxyHeaders = Record<string, string | string[] | undefined>;
+
+// Hop-by-hop headers (RFC 7230 §6.1) plus framing headers we manage ourselves.
+// Never forwarded downstream.
+export const STRIP_RESPONSE_HEADERS = new Set([
+  'connection', 'keep-alive', 'proxy-authenticate', 'proxy-authorization',
+  'te', 'trailer', 'transfer-encoding', 'upgrade',
+  // Framing controls — stripped so the proxy origin (which the webview frames)
+  // never inherits a "do not embed" from the upstream. For loopback that is the
+  // whole point; for a frameable remote there is nothing to strip anyway, and a
+  // refusing remote is diverted to an error page before we get here.
+  'x-frame-options', 'content-security-policy', 'content-security-policy-report-only',
+]);
+
+// The fixed, Dormouse-owned shim — like agent-browser's EDIT_SCRIPTS, never
+// user-supplied, so it is not an eval vector. Injected inline into served HTML
+// (loopback CSP is dropped, so an inline script runs). It reclaims ONLY the
+// reserved leader chord (dual-tap ⌘ / ⇧, matching handle-dual-tap.ts) and posts
+// it to the Wall; every other keystroke flows to the tool untouched. The focus
+// model needs no message channel — `document.hasFocus()` already stays true
+// while a descendant frame holds focus (resolves #2), and the Wall focuses the
+// frame element directly (resolves #3).
+export const IFRAME_SHIM = `(function(){
+  var P=window.parent;
+  if(!P||P===window)return;
+  function tap(s,e){
+    var now=Date.now(),side=e.location===1?'left':'right';
+    if(s.side==='left'&&side==='right'&&now-s.time<500){s.side=null;return true;}
+    s.side=side;s.time=now;return false;
+  }
+  function leader(){try{P.postMessage({__dormouse:'leader'},'*');}catch(e){}}
+  var cmd={side:null,time:0},shift={side:null,time:0};
+  addEventListener('keydown',function(e){
+    if(e.key==='Meta'){if(tap(cmd,e))leader();}
+    else if(e.key==='Shift'){if(tap(shift,e))leader();}
+  },true);
+})();`;
+
+// Drop any in-document CSP (loopback "relax CSP") and inject the shim before
+// </head> so it runs before the tool's own scripts. The response-header CSP is
+// stripped separately via STRIP_RESPONSE_HEADERS.
+export function instrumentHtml(body: string): string {
+  const html = body.replace(
+    /<meta[^>]+http-equiv=["']?content-security-policy["']?[^>]*>/gi,
+    '',
+  );
+  const shimTag = `<script>${IFRAME_SHIM}</script>`;
+  if (/<\/head>/i.test(html)) return html.replace(/<\/head>/i, `${shimTag}</head>`);
+  if (/<body[^>]*>/i.test(html)) return html.replace(/(<body[^>]*>)/i, `$1${shimTag}`);
+  return shimTag + html;
+}
+
+// A remote refuses framing if it sends any X-Frame-Options or a CSP
+// frame-ancestors that is not the permissive standalone `*`. Conservative on
+// purpose: when in doubt we divert to an error page rather than show a
+// guaranteed-blank frame.
+export function refusesFraming(headers: ProxyHeaders): boolean {
+  if (headers['x-frame-options']) return true;
+  const csp = headers['content-security-policy'];
+  const policies = Array.isArray(csp) ? csp : csp ? [csp] : [];
+  return policies.some((policy) => hasRestrictiveFrameAncestors(policy));
+}
+
+export function hasRestrictiveFrameAncestors(policy: string): boolean {
+  const directives = policy.split(';');
+  for (const directive of directives) {
+    const parts = directive.trim().split(/\s+/).filter(Boolean);
+    if (parts.length === 0 || parts[0].toLowerCase() !== 'frame-ancestors') continue;
+    const sources = parts.slice(1);
+    if (!sources.includes('*')) return true;
+  }
+  return false;
+}
+
+export function isLoopbackHost(hostname: string): boolean {
+  const h = hostname.replace(/^\[|\]$/g, '').toLowerCase();
+  return h === 'localhost' || h === '127.0.0.1' || h === '::1' || h.startsWith('127.');
+}
+
+export function isBlockedAddress(hostname: string): boolean {
+  const h = hostname.replace(/^\[|\]$/g, '').toLowerCase();
+  // IPv4 link-local / cloud metadata (169.254.0.0/16, incl. 169.254.169.254).
+  if (/^169\.254\./.test(h)) return true;
+  // IPv6 link-local (fe80::/10).
+  if (/^fe[89ab][0-9a-f]:/.test(h)) return true;
+  return false;
+}
+
+// --- Served error / diagnostic pages ----------------------------------------
+
+export interface ErrorPage {
+  title: string;
+  message: string;
+  hint?: string;
+}
+
+export function frameRefusedPage(upstream: URL): ErrorPage {
+  return {
+    title: `${upstream.host} refuses to be embedded`,
+    message: `${upstream.host} sends a frame-blocking header (X-Frame-Options or CSP frame-ancestors), so it can’t be shown in an iframe surface.`,
+    hint: `dor ab open ${upstream.href}`,
+  };
+}
+
+export function unreachablePage(upstream: URL, detail: string): ErrorPage {
+  return {
+    title: `Nothing responding at ${upstream.host}`,
+    message: `Dormouse couldn’t reach ${upstream.href} (${detail}). Is the dev server running?`,
+  };
+}
+
+export function escapeHtml(value: string): string {
+  return value
+    .replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;')
+    .replace(/"/g, '&quot;');
+}
+
+export function errorPageHtml(page: ErrorPage): string {
+  const hint = page.hint
+    ? `<p class="hint">Try <code>${escapeHtml(page.hint)}</code></p>`
+    : '';
+  return `<!doctype html><html><head><meta charset="utf-8">
+<style>
+  :root { color-scheme: dark; }
+  html, body { height: 100%; margin: 0; }
+  body { display: flex; align-items: center; justify-content: center;
+    background: #14161a; color: #c9ced6;
+    font: 14px/1.5 -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif; }
+  .card { max-width: 34rem; padding: 1.5rem 2rem; text-align: center; }
+  h1 { margin: 0 0 .5rem; font-size: 1.05rem; font-weight: 600; color: #e7ebf1; }
+  p { margin: .5rem 0; }
+  .hint { margin-top: 1rem; }
+  code { background: #20242b; border-radius: 4px; padding: .15rem .4rem;
+    font-family: ui-monospace, SFMono-Regular, Menlo, monospace; color: #e7ebf1; }
+</style></head>
+<body><div class="card">
+  <h1>${escapeHtml(page.title)}</h1>
+  <p>${escapeHtml(page.message)}</p>
+  ${hint}
+</div></body></html>`;
+}
